Fix x64 compatibility issues in hash functions and printf formats
Critical hash function fixes:
- ivu_min_hash: Change hash_index parameter from int* to intp* to hash
  all 64 bits of pointers. Add upper 32-bit mixing on 64-bit platforms.
  Update callers (add, remove, is_elem) to use intp* casts.
- ivu_vhash: Change fast_hash_index from int to intp parameter.
  Use portable bit shifting (sizeof(intp)*4 instead of hardcoded 16).
- ivu_set: Cast element to intp instead of long for hash_index call,
  since long is only 32-bit on Windows x64.

Printf format fixes for Windows x64 (where long is 32-bit but intp is 64-bit):
- ivp_object_polygon_tetra.cxx: %lX -> %X with (int) cast
- ivp_calc_next_psi_solver.cxx: %lx -> %x with (int) cast
- ivp_core.cxx: %lx -> %x with (int) cast
- ivp_friction_gaps.cxx: %lx -> %zx for intp values

These fixes prevent pointer truncation bugs that would cause hash collisions
and incorrect lookups when running on 64-bit systems.

+++ ivp_utility/ivu_min_hash.hxx
@@ -29,7 +29,7 @@ extern unsigned int IVP_Hash_crctab[];
 class IVP_U_Min_Hash
 {
     friend class IVP_U_Min_Hash_Enumerator;
-    inline int hash_index(const int *key) const;
+    inline int hash_index(const intp *key) const;
     unsigned int size;
     IVP_U_Min_Hash_Elem **stadel;
     IVP_U_Min_Hash_Elem **min_per_array_pos;
